refactor: simplify project struct
