+++ project/plugin.sbt
@@ -1,2 +1,4 @@
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.1")
+// https://github.com/scala-ide/scalariform
+// https://github.com/sbt/sbt-scalariform
 addSbtPlugin("org.scalariform" % "sbt-scalariform" % "1.8.2")
