+++ readme.md
@@ -1,6 +1,8 @@
-# [WIP] kafka t_metadata
+# [WIP] KTM
+KTM stands for Kafka Topic Metadata.
+
 REST api is aiming to expand kafka-topics metadata, by storing it in another topic.
 
 ## TODO
 1. Kafka Streams app in actor
-2. Supervisor
+2. ActorObserver
