+++ src/main/scala/no/sysco/middleware/ktm/infrastructure/Topics.scala
@@ -1,0 +1,6 @@
+package no.sysco.middleware.ktm.infrastructure
+
+object Topics {
+  val METADATA: String = "__topics"
+  val METADATA_STORAGE: String = s"${Topics.METADATA}_storage"
+}
