+++ src/test/scala/no/sysco/middleware/ktm/BootTest.scala
@@ -1,0 +1,5 @@
+package no.sysco.middleware.ktm
+
+class BootTest {
+
+}
