+++ src/test/scala/no/sysco/middleware/ktm/utils/Utils.scala
@@ -1,0 +1,31 @@
+package no.sysco.middleware.ktm.utils
+
+
+
+import no.sysco.middleware.ktm.rest.{Team, TopicMetadata, TopicVendor}
+
+import scala.io.Source
+
+object Utils {
+
+  def jsonFromFile(filename: String): String = Source.fromFile(filename).mkString
+  def metadataPojo(withSla: Boolean): TopicMetadata = {
+    TopicMetadata(
+      "some name",
+      "some description",
+      List(
+        Team("team name1", "department name1"),
+        Team("team name2", "department name2")
+      ),
+      TopicVendor("STATTNETT"),
+      "XML",
+      "some scope",
+      "some config",
+      withSla match {
+        case true => Option("some sla")
+        case false => Option.empty
+      }
+    )
+  }
+
+}
