--- tmetadata/src/main/scala/no/sysco/middleware/kafka/tmetadata/infrastructure/Topics.scala
@@ -1,6 +1,0 @@
-package no.sysco.middleware.kafka.tmetadata.infrastructure
-
-object Topics {
-  val METADATA: String = "__topics"
-  val METADATA_STORAGE: String = s"${Topics.METADATA}_storage"
-}
