--- tmetadata/src/main/scala/no/sysco/middleware/kafka/tmetadata/rest/routes/HealthRoutes.scala
@@ -1,16 +1,0 @@
-package no.sysco.middleware.kafka.tmetadata.rest.routes
-
-import akka.http.scaladsl.model.{ ContentTypes, HttpEntity }
-import akka.http.scaladsl.server.Directives.{ complete, get, path, pathEndOrSingleSlash }
-
-trait HealthRoutes {
-
-  val healthRoutes = path("health") {
-    pathEndOrSingleSlash {
-      get {
-        complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, "OK"))
-      }
-    }
-  }
-
-}
