--- tmetadata/src/test/scala/no/sysco/middleware/kafka/tmetadata/BootTest.scala
@@ -1,5 +1,0 @@
-package no.sysco.middleware.kafka.tmetadata
-
-class BootTest {
-
-}
